% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_coordinate_convert.R
\docType{data}
\name{ee_constant}
\alias{ee_constant}
\title{eccentricity}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
ee_constant
}
\description{
eccentricity
}
\keyword{datasets}
