% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocn_datasets.R
\name{list_geocn}
\alias{list_geocn}
\title{List all data sets available in the geocn package}
\usage{
list_geocn()
}
\value{
A tibble
}
\description{
Returns a tibble with all datasets available in the geocn package.
}
\examples{
list_geocn()

}
