% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_albersproj.R
\name{load_cn_alberproj}
\alias{load_cn_alberproj}
\title{Loading the commonly used Albers Lambert projection in China region.}
\usage{
load_cn_alberproj(output = "sf")
}
\arguments{
\item{output}{(optional) Output format, must be \code{sf} or \code{terra}. Default is \code{sf}.}
}
\value{
An Available Projection CRS.
}
\description{
Loading the commonly used Albers Lambert projection in China region.
}
\examples{
load_cn_alberproj()

}
