% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_boundary.R
\name{load_cn_border}
\alias{load_cn_border}
\title{Loading China's Land Border Line and the 10-dash line of the South China Sea}
\usage{
load_cn_border()
}
\value{
An sf object
}
\description{
Loading China's Land Border Line and the 10-dash line of the South China Sea
}
\examples{
cn_border = load_cn_border()
cn_border

}
