% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_boundary.R
\name{load_cn_coastline}
\alias{load_cn_coastline}
\title{Loading Coastline of China}
\usage{
load_cn_coastline()
}
\value{
An sf object
}
\description{
Loading Coastline of China
}
\examples{
cn_coastline = load_cn_coastline()
cn_coastline

}
