% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_boundary.R
\name{load_cn_landborder}
\alias{load_cn_landborder}
\title{Loading China's Land Border}
\usage{
load_cn_landborder()
}
\value{
An sf object
}
\description{
Loading China's Land Border
}
\examples{
cn_landborder = load_cn_landborder()
cn_landborder

}
