% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_boundary.R
\name{load_cn_landcoast}
\alias{load_cn_landcoast}
\title{Loading China's Land Border and Coastline}
\usage{
load_cn_landcoast()
}
\value{
An sf object
}
\description{
Loading China's Land Border and Coastline
}
\examples{
cn_landcoast = load_cn_landcoast()
cn_landcoast

}
