% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_boundary.R
\name{load_cn_tenline}
\alias{load_cn_tenline}
\title{Loading the 10-dash line of the South China Sea}
\usage{
load_cn_tenline()
}
\value{
An sf object
}
\description{
Loading the 10-dash line of the South China Sea
}
\examples{
cn_tenline = load_cn_tenline()
cn_tenline

}
