% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_typical_region.R
\name{load_loess_plateau}
\alias{load_loess_plateau}
\title{Loading the Boundary Polygon Data of Loess Plateau}
\usage{
load_loess_plateau()
}
\value{
An sf object
}
\description{
Loading the Boundary Polygon Data of Loess Plateau
}
\examples{
load_loess_plateau()

}
