% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_administrative.R
\name{load_province_cn}
\alias{load_province_cn}
\title{Loading Chinese Provincial Administrative Divisional Polygon Data}
\usage{
load_province_cn(simplify = TRUE, keep = 0.05, keep_shape = TRUE, ...)
}
\arguments{
\item{simplify}{(optional) Whether to simplify loading sf objects (default TRUE)}

\item{keep}{(optional) Proportion of points to retain (0-1; default 0.05)}

\item{keep_shape}{(optional)    Prevent small polygon features from disappearing at high simplification (default TRUE)}

\item{...}{(optional) Other arguments passed to \code{rmapshaper::ms_simplify()}}
}
\value{
Whether to simplify loading sf objects
}
\description{
Loading Chinese Provincial Administrative Divisional Polygon Data
}
\examples{
library(sf)
province = load_province_cn()
province

}
