% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_typical_region.R
\name{load_tarim_basin}
\alias{load_tarim_basin}
\title{Loading the Boundary Polygon Data of Tarim River Basin}
\usage{
load_tarim_basin()
}
\value{
An sf object
}
\description{
Loading the Boundary Polygon Data of Tarim River Basin
}
\examples{
load_tarim_basin()

}
