% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_typical_region.R
\name{load_tibetan_plateau}
\alias{load_tibetan_plateau}
\title{Loading the Boundary Polygon Data of Tibetan Plateau}
\usage{
load_tibetan_plateau()
}
\value{
An sf object
}
\description{
Loading the Boundary Polygon Data of Tibetan Plateau
}
\examples{
load_tibetan_plateau()

}
