% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_typical_region.R
\name{load_weihe_basin}
\alias{load_weihe_basin}
\title{Loading the Boundary Polygon Data of Weihe River Basin}
\usage{
load_weihe_basin()
}
\value{
An sf object
}
\description{
Loading the Boundary Polygon Data of Weihe River Basin
}
\examples{
load_weihe_basin()

}
