% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_vector.R
\name{load_world_coastline}
\alias{load_world_coastline}
\title{Loading Global Coastlines}
\usage{
load_world_coastline()
}
\value{
An sf object
}
\description{
Loading Global Coastlines
}
\examples{
load_world_coastline()

}
