% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_vector.R
\name{load_world_continent}
\alias{load_world_continent}
\title{Loading Global Continents}
\usage{
load_world_continent()
}
\value{
An sf object
}
\description{
Loading Global Continents
}
\examples{
load_world_continent()

}
