% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_vector.R
\name{load_world_country}
\alias{load_world_country}
\title{Load Global Country Boundaries}
\usage{
load_world_country(center = "west")
}
\arguments{
\item{center}{(optional) Center must be \code{west} or \code{east}. Default is \code{west}.}
}
\value{
An sf object
}
\description{
Load Global Country Boundaries
}
\details{
When the \code{center} parameter is set to \code{west}, the map center is the Atlantic Ocean;
and when \code{center} is \verb{east,} the map center is the Pacific Ocean.
}
\examples{
load_world_country()

}
