% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_typical_region.R
\name{load_yellow_river_basin}
\alias{load_yellow_river_basin}
\title{Loading the Boundary Polygon Data of Yellow River Basin}
\usage{
load_yellow_river_basin()
}
\value{
An sf object
}
\description{
Loading the Boundary Polygon Data of Yellow River Basin
}
\examples{
load_yellow_river_basin()

}
