% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_coordinate_convert.R
\docType{data}
\name{pf_constant}
\alias{pf_constant}
\title{projection factor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
pf_constant
}
\description{
projection factor
}
\keyword{datasets}
